+++ package.json
@@ -110,24 +110,24 @@
     "@semantic-release/commit-analyzer": "^9.0.2",
     "@semantic-release/github": "^8.0.7",
     "@semantic-release/release-notes-generator": "^10.0.3",
-    "@storybook/addon-a11y": "^7.0.0-rc.0",
-    "@storybook/addon-actions": "^7.0.0-rc.0",
-    "@storybook/addon-controls": "^7.0.0-rc.0",
-    "@storybook/addon-docs": "^7.0.0-rc.0",
-    "@storybook/addon-essentials": "^7.0.0-rc.0",
-    "@storybook/addon-interactions": "^7.0.0-rc.0",
-    "@storybook/addon-links": "^7.0.0-rc.0",
+    "@storybook/addon-a11y": "^7.0.0-rc.1",
+    "@storybook/addon-actions": "^7.0.0-rc.1",
+    "@storybook/addon-controls": "^7.0.0-rc.1",
+    "@storybook/addon-docs": "^7.0.0-rc.1",
+    "@storybook/addon-essentials": "^7.0.0-rc.1",
+    "@storybook/addon-interactions": "^7.0.0-rc.1",
+    "@storybook/addon-links": "^7.0.0-rc.1",
     "@storybook/addon-postcss": "^2.0.0",
-    "@storybook/addons": "^7.0.0-rc.0",
-    "@storybook/api": "^7.0.0-rc.0",
-    "@storybook/blocks": "^7.0.0-rc.0",
-    "@storybook/components": "^7.0.0-rc.0",
-    "@storybook/core-events": "^7.0.0-rc.0",
-    "@storybook/manager-api": "^7.0.0-rc.0",
-    "@storybook/react": "^7.0.0-rc.0",
-    "@storybook/react-webpack5": "^7.0.0-rc.0",
+    "@storybook/addons": "^7.0.0-rc.1",
+    "@storybook/api": "^7.0.0-rc.1",
+    "@storybook/blocks": "^7.0.0-rc.1",
+    "@storybook/components": "^7.0.0-rc.1",
+    "@storybook/core-events": "^7.0.0-rc.1",
+    "@storybook/manager-api": "^7.0.0-rc.1",
+    "@storybook/react": "^7.0.0-rc.1",
+    "@storybook/react-webpack5": "^7.0.0-rc.1",
     "@storybook/testing-library": "^0.0.14-next.1",
-    "@storybook/theming": "^7.0.0-rc.0",
+    "@storybook/theming": "^7.0.0-rc.1",
     "@testing-library/dom": "^9.0.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^14.0.0",
@@ -186,7 +186,7 @@
     "rimraf": "^4.3.0",
     "semantic-release": "^20.1.0",
     "semantic-release-export-data": "^1.0.1",
-    "storybook": "^7.0.0-rc.0",
+    "storybook": "^7.0.0-rc.1",
     "storybook-addon-themes": "^6.1.0",
     "storybook-dark-mode-v7": "^3.0.0-alpha.0",
     "stylelint": "^15.1.0",
